Move TrackDataset into separate module
