+++ OTAnalytics/application/datastore.py
@@ -22,7 +22,8 @@ from OTAnalytics.domain.section import (
     SectionListObserver,
     SectionRepository,
 )
-from OTAnalytics.domain.track import TrackDataset, TrackId, TrackImage
+from OTAnalytics.domain.track import TrackId, TrackImage
+from OTAnalytics.domain.track_dataset import TrackDataset
 from OTAnalytics.domain.track_repository import (
     TrackFileRepository,
     TrackListObserver,
