+++ OTAnalytics/application/use_cases/create_intersection_events.py
@@ -14,7 +14,7 @@ from OTAnalytics.domain.geometry import (
 )
 from OTAnalytics.domain.intersect import Intersector, IntersectParallelizationStrategy
 from OTAnalytics.domain.section import Area, LineSection, Section
-from OTAnalytics.domain.track import TrackDataset
+from OTAnalytics.domain.track_dataset import TrackDataset
 from OTAnalytics.domain.types import EventType
