+++ OTAnalytics/application/use_cases/track_repository.py
@@ -2,7 +2,8 @@ from pathlib import Path
 from typing import Iterable
 
 from OTAnalytics.application.logger import logger
-from OTAnalytics.domain.track import Track, TrackDataset, TrackId
+from OTAnalytics.domain.track import Track, TrackId
+from OTAnalytics.domain.track_dataset import TrackDataset
 from OTAnalytics.domain.track_repository import (
     RemoveMultipleTracksError,
     TrackFileRepository,
