+++ OTAnalytics/domain/track_dataset.py
@@ -1,0 +1,136 @@
+from abc import ABC, abstractmethod
+from dataclasses import dataclass
+from typing import Any, Callable, Iterable, Iterator, Optional, Sequence
+
+from OTAnalytics.domain.geometry import RelativeOffsetCoordinate
+from OTAnalytics.domain.section import Section, SectionId
+from OTAnalytics.domain.track import Track, TrackId
+
+
+@dataclass
+class IntersectionPoint:
+    index: int
+
+
+class TrackDataset(ABC):
+    def __iter__(self) -> Iterator[Track]:
+        yield from self.as_list()
+
+    @abstractmethod
+    def add_all(self, other: Iterable[Track]) -> "TrackDataset":
+        raise NotImplementedError
+
+    @abstractmethod
+    def get_all_ids(self) -> Iterable[TrackId]:
+        raise NotImplementedError
+
+    @abstractmethod
+    def get_for(self, id: TrackId) -> Optional[Track]:
+        """
+        Retrieve a track for the given id.
+
+        Args:
+            id (TrackId): id to search for
+
+        Returns:
+            Optional[Track]: track if it exists
+        """
+        raise NotImplementedError
+
+    @abstractmethod
+    def remove(self, track_id: TrackId) -> "TrackDataset":
+        raise NotImplementedError
+
+    @abstractmethod
+    def clear(self) -> "TrackDataset":
+        """
+        Return an empty version of the current TrackDataset.
+        """
+        raise NotImplementedError
+
+    @abstractmethod
+    def as_list(self) -> list[Track]:
+        raise NotImplementedError
+
+    @abstractmethod
+    def intersecting_tracks(
+        self, sections: list[Section], offset: RelativeOffsetCoordinate
+    ) -> set[TrackId]:
+        """Return a set of tracks intersecting a set of sections.
+
+        Args:
+            sections (list[Section]): the list of sections to intersect.
+            offset (RelativeOffsetCoordinate): the offset to be applied to the tracks.
+
+        Returns:
+            set[TrackId]: the track ids intersecting the given sections.
+        """
+        raise NotImplementedError
+
+    @abstractmethod
+    def intersection_points(
+        self, sections: list[Section], offset: RelativeOffsetCoordinate
+    ) -> dict[TrackId, list[tuple[SectionId, IntersectionPoint]]]:
+        """
+        Return the intersection points resulting from the tracks and the
+        given sections.
+
+        Args:
+            sections (list[Section]): the sections to intersect with.
+            offset (RelativeOffsetCoordinate): the offset to be applied to the tracks.
+
+        Returns:
+            dict[TrackId, list[tuple[SectionId]]]: the intersection points.
+        """
+        raise NotImplementedError
+
+    @abstractmethod
+    def contained_by_sections(
+        self, sections: list[Section], offset: RelativeOffsetCoordinate
+    ) -> dict[TrackId, list[tuple[SectionId, list[bool]]]]:
+        """Return whether track coordinates are contained by the given sections.
+
+        Args:
+             sections (list[Section]): the sections.
+             offset (RelativeOffsetCoordinate): the offset to be applied to the tracks.
+
+        Returns:
+            dict[TrackId, list[tuple[SectionId, list[bool]]]]: boolean mask of track
+                coordinates contained by given sections.
+        """
+        raise NotImplementedError
+
+    @abstractmethod
+    def split(self, chunks: int) -> Sequence["TrackDataset"]:
+        raise NotImplementedError
+
+    @abstractmethod
+    def __len__(self) -> int:
+        """Number of tracks in the dataset."""
+        raise NotImplementedError
+
+    @abstractmethod
+    def filter_by_min_detection_length(self, length: int) -> "TrackDataset":
+        """Filter tracks by the minimum length of detections.
+
+        Args:
+            length (int): minimum number detections a track should have.
+
+        Returns:
+             TrackDataset: the filtered dataset.
+        """
+        raise NotImplementedError
+
+    @abstractmethod
+    def calculate_geometries_for(
+        self, offsets: Iterable[RelativeOffsetCoordinate]
+    ) -> None:
+        raise NotImplementedError
+
+    @abstractmethod
+    def apply_to_first_segments(self, consumer: Callable[[Any], None]) -> None:
+        raise NotImplementedError
+
+    @abstractmethod
+    def apply_to_last_segments(self, consumer: Callable[[Any], None]) -> None:
+        raise NotImplementedError
