+++ OTAnalytics/plugin_datastore/python_track_store.py
@@ -17,14 +17,13 @@ from OTAnalytics.domain.section import Section, SectionId
 from OTAnalytics.domain.track import (
     TRACK_GEOMETRY_FACTORY,
     Detection,
-    IntersectionPoint,
     Track,
     TrackClassificationCalculator,
-    TrackDataset,
     TrackGeometryDataset,
     TrackHasNoDetectionError,
     TrackId,
 )
+from OTAnalytics.domain.track_dataset import IntersectionPoint, TrackDataset
 from OTAnalytics.domain.types import EventType
 from OTAnalytics.plugin_datastore.track_geometry_store.pygeos_store import (
     PygeosTrackGeometryDataset,
