+++ OTAnalytics/plugin_intersect_parallelization/multiprocessing.py
@@ -7,7 +7,7 @@ from OTAnalytics.application.logger import logger
 from OTAnalytics.domain.event import Event
 from OTAnalytics.domain.intersect import IntersectParallelizationStrategy
 from OTAnalytics.domain.section import Section
-from OTAnalytics.domain.track import TrackDataset
+from OTAnalytics.domain.track_dataset import TrackDataset
 
 
 class MultiprocessingIntersectParallelization(IntersectParallelizationStrategy):
