+++ OTAnalytics/plugin_intersect_parallelization/sequential.py
@@ -3,7 +3,7 @@ from typing import Callable, Iterable, Sequence
 from OTAnalytics.domain.event import Event
 from OTAnalytics.domain.intersect import IntersectParallelizationStrategy
 from OTAnalytics.domain.section import Section
-from OTAnalytics.domain.track import TrackDataset
+from OTAnalytics.domain.track_dataset import TrackDataset
 
 
 class SequentialIntersect(IntersectParallelizationStrategy):
