+++ OTAnalytics/plugin_parser/otvision_parser.py
@@ -38,11 +38,11 @@ from OTAnalytics.domain.track import (
     Detection,
     Track,
     TrackClassificationCalculator,
-    TrackDataset,
     TrackHasNoDetectionError,
     TrackId,
     TrackImage,
 )
+from OTAnalytics.domain.track_dataset import TrackDataset
 from OTAnalytics.domain.track_repository import TrackRepository
 from OTAnalytics.domain.video import PATH, SimpleVideo, Video, VideoReader
 from OTAnalytics.plugin_datastore.python_track_store import (
