+++ OTAnalytics/plugin_parser/pandas_parser.py
@@ -5,7 +5,8 @@ from pandas import DataFrame
 
 from OTAnalytics.application.logger import logger
 from OTAnalytics.domain import track
-from OTAnalytics.domain.track import TRACK_GEOMETRY_FACTORY, TrackDataset, TrackId
+from OTAnalytics.domain.track import TRACK_GEOMETRY_FACTORY, TrackId
+from OTAnalytics.domain.track_dataset import TrackDataset
 from OTAnalytics.plugin_datastore.track_store import (
     PandasTrackClassificationCalculator,
     PandasTrackDataset,
