+++ tests/OTAnalytics/application/test_eventlist.py
@@ -12,7 +12,8 @@ from OTAnalytics.domain.geometry import (
     RelativeOffsetCoordinate,
 )
 from OTAnalytics.domain.section import LineSection, SectionId
-from OTAnalytics.domain.track import Detection, Track, TrackDataset, TrackId
+from OTAnalytics.domain.track import Detection, Track, TrackId
+from OTAnalytics.domain.track_dataset import TrackDataset
 from OTAnalytics.plugin_datastore.python_track_store import PythonDetection, PythonTrack
