+++ tests/OTAnalytics/application/use_cases/test_track_repository.py
@@ -15,7 +15,8 @@ from OTAnalytics.application.use_cases.track_repository import (
     RemoveTracks,
     TrackRepositorySize,
 )
-from OTAnalytics.domain.track import Track, TrackDataset, TrackId
+from OTAnalytics.domain.track import Track, TrackId
+from OTAnalytics.domain.track_dataset import TrackDataset
 from OTAnalytics.domain.track_repository import TrackFileRepository, TrackRepository
