+++ tests/OTAnalytics/domain/test_track.py
@@ -3,7 +3,8 @@ from unittest.mock import MagicMock, Mock, call
 
 import pytest
 
-from OTAnalytics.domain.track import Track, TrackDataset, TrackId
+from OTAnalytics.domain.track import Track, TrackId
+from OTAnalytics.domain.track_dataset import TrackDataset
 from OTAnalytics.domain.track_repository import (
     TrackFileRepository,
     TrackListObserver,
