+++ tests/OTAnalytics/plugin_datastore/test_track_store.py
@@ -14,10 +14,10 @@ from OTAnalytics.domain.geometry import (
 from OTAnalytics.domain.track import (
     TRACK_GEOMETRY_FACTORY,
     Track,
-    TrackDataset,
     TrackGeometryDataset,
     TrackId,
 )
+from OTAnalytics.domain.track_dataset import TrackDataset
 from OTAnalytics.domain.types import EventType
 from OTAnalytics.plugin_datastore.python_track_store import (
     PythonTrack,
